Add cli option to change scaling of the signature
Before this change, the signature was taken as is from the signature directory in the size it was scanned. Yet sometimes the bureaucracy folks come up with forms with little fields for the signature. Then one needs to use this scaling feature, e.g. "-S 80" to place the signature of 80% of the original scanned template signature size.